configure auth

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -58,13 +59,13 @@ public class AuctionsController : ControllerBase
         return dataToReturn;
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<List<AuctionDto>>> CreateAuction(AuctionCreateDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
 
-        // TODO: add current user as seller
-        auction.Seller = "test";
+        auction.Seller = User.Identity!.Name;
 
         _context.Auctions.Add(auction);
 
@@ -80,6 +81,7 @@ public class AuctionsController : ControllerBase
             new { id = auction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, AuctionUpdateDto auctionDto)
     {
@@ -88,6 +90,8 @@ public class AuctionsController : ControllerBase
             .FirstOrDefaultAsync(x => x.Id == id);
 
         if (auction == null) return NotFound();
+        
+        if(auction.Seller != User.Identity!.Name) return Forbid();
 
         auction.Item.Make = auctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = auctionDto.Model ?? auction.Item.Model;
@@ -104,6 +108,7 @@ public class AuctionsController : ControllerBase
         return BadRequest("Could not update the auction");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -111,7 +116,7 @@ public class AuctionsController : ControllerBase
 
         if (auction == null) return NotFound();
 
-        //TODO - check seller == username
+        if(auction.Seller != User.Identity!.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
